Move AAC Frechet mean to frechet mean for consistency
